backend/requirements.txt
flask
flask_cors
flask_jwt_extended
flask_sqlalchemy
pymysql
cryptography
python-dotenv
